refactor(cli): extract main.py functions into utils modules
Move 17 functions out of main.py into their logical domains:
- utils/config.py: resolve_config, validate_cli_arguments, parsers
- utils/output.py: display, storage, path resolution (9 functions)
- utils/diff.py: handle_diff_output, resolve_diff_flags
- utils/snapshot.py: handle_snapshot (refactored with guard clauses)
- utils/ignored.py: handle_report_ignored

main.py is now a pure orchestrator with only _version_callback and
main(). Module-level globals (TOML_CONFIG, INVOCATION_PATH) are passed
as parameters instead of referenced implicitly.
