+++ complexipy/utils/output.py
@@ -1,12 +1,15 @@
 from __future__ import annotations
 
+import os
+import platform
 from typing import (
     Dict,
-    List,  # It's important to use this to make it compatible with python 3.8, don't remove it
+    List,
     Optional,
     Tuple,
 )
 
+import typer
 from rich.console import Console
 from rich.text import Text
 
@@ -15,9 +18,266 @@ from complexipy._complexipy import (
     FunctionComplexity,
 )
 from complexipy.types import (
+    ColorTypes,
+    OutputFormat,
     Sort,
 )
+from complexipy.utils.cache import remember_previous_functions
+from complexipy.utils.constants import (
+    DEFAULT_OUTPUT_FILENAMES,
+    LEGACY_OUTPUT_CONFIG_KEYS,
+    LEGACY_OUTPUT_FLAGS,
+)
 from complexipy.utils.dataclasses import FileEntry, FunctionRow
+
+
+def handle_console_settings(
+    color: ColorTypes, quiet: bool, plain: bool = False
+) -> Console:
+    if plain:
+        return Console(color_system=None, highlight=False)
+
+    if color == ColorTypes.no:
+        console = Console(color_system=None)
+    elif color == ColorTypes.yes:
+        console = Console(color_system="standard")
+    else:
+        console = Console(color_system="auto")
+
+    if not quiet:
+        if platform.system() == "Windows":
+            console.rule("complexipy")
+        else:
+            console.rule(":octopus: complexipy")
+
+    return console
+
+
+def handle_display(
+    console: Console,
+    files_complexities: List[FileComplexity],
+    paths: List[str],
+    failed: bool,
+    sort: Sort,
+    ignore_complexity: bool,
+    max_complexity_allowed: int,
+    active_snapshot_map: Optional[Dict],
+    quiet: bool,
+    plain: bool,
+    invocation_path: str,
+    top: Optional[int] = None,
+    suggest_refactors: bool = False,
+) -> bool:
+    if files_complexities:
+        previous_functions = remember_previous_functions(
+            invocation_path, paths, files_complexities
+        )
+    else:
+        previous_functions = None
+
+    if quiet:
+        return has_success_functions(
+            files_complexities, max_complexity_allowed, active_snapshot_map
+        )
+
+    effective_sort = Sort.desc if top is not None else sort
+    has_success = output_summary(
+        console,
+        files_complexities,
+        failed,
+        effective_sort,
+        ignore_complexity,
+        max_complexity_allowed,
+        previous_functions,
+        active_snapshot_map,
+        plain,
+        top,
+        suggest_refactors,
+    )
+    return has_success
+
+
+def handle_results_storage(
+    console: Console,
+    output_formats: List[OutputFormat],
+    output: Optional[str],
+    files_complexities: List[FileComplexity],
+    sort: str,
+    show_details: bool,
+    max_complexity: int,
+    invocation_path: str,
+) -> None:
+    output_paths = resolve_output_paths(output_formats, output, invocation_path)
+
+    for output_format in output_formats:
+        output_path = output_paths[output_format]
+
+        if output_format == OutputFormat.csv:
+            from complexipy.utils.csv import store_csv
+
+            store_csv(
+                output_path,
+                files_complexities,
+                sort,
+                show_details,
+                max_complexity,
+            )
+        elif output_format == OutputFormat.json:
+            from complexipy.utils.json import store_json
+
+            store_json(
+                output_path,
+                files_complexities,
+                show_details,
+                max_complexity,
+            )
+        elif output_format == OutputFormat.gitlab:
+            from complexipy.utils.gitlab import store_gitlab
+
+            store_gitlab(
+                output_path,
+                files_complexities,
+                max_complexity,
+            )
+        elif output_format == OutputFormat.sarif:
+            from complexipy.utils.sarif import store_sarif
+
+            store_sarif(
+                output_path,
+                files_complexities,
+                max_complexity,
+            )
+
+        console.print(f"Results saved at {output_path}")
+
+
+def emit_deprecated_output_warnings(
+    console: Console,
+    legacy_cli_output_flags: Dict[OutputFormat, Optional[bool]],
+    toml_config,
+) -> None:
+    for output_format, flag_name in LEGACY_OUTPUT_FLAGS.items():
+        if legacy_cli_output_flags[output_format]:
+            console.print(
+                f"[yellow]Deprecated:[/yellow] {flag_name} will be removed "
+                f"in a future release. Use `--output-format "
+                f"{output_format.value}` instead."
+            )
+
+    if toml_config is None:
+        return
+
+    for output_format, config_key in LEGACY_OUTPUT_CONFIG_KEYS.items():
+        if bool(toml_config.get(config_key, False)):
+            console.print(
+                f"[yellow]Deprecated:[/yellow] `{config_key}` in TOML will "
+                f"be removed in a future release. Use `output-format = "
+                f'["{output_format.value}"]` instead.'
+            )
+
+
+def resolve_output_formats(
+    output_format_values: List[str],
+    legacy_cli_output_flags: Dict[OutputFormat, Optional[bool]],
+    toml_config,
+) -> List[OutputFormat]:
+    output_formats = []
+
+    for value in output_format_values:
+        try:
+            normalized = OutputFormat(value)
+        except ValueError as exc:
+            valid_values = ", ".join(
+                available.value for available in OutputFormat
+            )
+            raise typer.BadParameter(
+                f"Invalid output format '{value}'. Expected one of: "
+                f"{valid_values}."
+            ) from exc
+
+        if normalized not in output_formats:
+            output_formats.append(normalized)
+
+    for output_format in OutputFormat:
+        if legacy_cli_output_flags[output_format]:
+            output_formats.append(output_format)
+            continue
+
+        config_key = LEGACY_OUTPUT_CONFIG_KEYS[output_format]
+        if toml_config is not None and bool(toml_config.get(config_key, False)):
+            output_formats.append(output_format)
+
+    return list(dict.fromkeys(output_formats))
+
+
+def resolve_output_paths(
+    output_formats: List[OutputFormat],
+    output: Optional[str],
+    invocation_path: str,
+) -> Dict[OutputFormat, str]:
+    if not output_formats:
+        return {}
+
+    if output is None:
+        return build_output_paths(invocation_path, output_formats)
+
+    if output == "-":
+        raise typer.BadParameter(
+            "Writing machine-readable output to stdout is not supported."
+        )
+
+    destination = os.path.abspath(output)
+    is_directory_hint = is_directory_output_hint(output)
+
+    if len(output_formats) > 1:
+        ensure_directory_destination(destination, is_directory_hint)
+        return build_output_paths(destination, output_formats)
+
+    output_format = output_formats[0]
+    if os.path.isdir(destination) or is_directory_hint:
+        os.makedirs(destination, exist_ok=True)
+        return build_output_paths(destination, [output_format])
+
+    parent_dir = os.path.dirname(destination)
+    if parent_dir:
+        os.makedirs(parent_dir, exist_ok=True)
+
+    return {output_format: destination}
+
+
+def build_output_paths(
+    destination: str, output_formats: List[OutputFormat]
+) -> Dict[OutputFormat, str]:
+    return {
+        output_format: os.path.join(
+            destination,
+            DEFAULT_OUTPUT_FILENAMES[output_format],
+        )
+        for output_format in output_formats
+    }
+
+
+def is_directory_output_hint(output: str) -> bool:
+    return output.endswith(os.sep) or (
+        os.altsep is not None and output.endswith(os.altsep)
+    )
+
+
+def ensure_directory_destination(
+    destination: str, is_directory_hint: bool
+) -> None:
+    if os.path.exists(destination) and not os.path.isdir(destination):
+        raise typer.BadParameter(
+            "When multiple output formats are selected, --output "
+            "must point to a directory."
+        )
+    elif not is_directory_hint:
+        raise typer.BadParameter(
+            "When multiple output formats are selected, --output must "
+            "point to a directory or end with a path separator."
+        )
+
+    os.makedirs(destination, exist_ok=True)
 
 
 def output_summary(
